Authorithy checks added (#135)
* table authorization group added

* authorization package added

* abapDoc added to auth checker class

* abaplint fixes

+++ src/zamp_store/zamp_clear_zamp_store.prog.abap
@@ -5,4 +5,8 @@
 *&---------------------------------------------------------------------*
 REPORT zamp_clear_zamp_store.
 
-DELETE FROM zamp_store.
+IF NEW zcl_amp_auth_checker( )->is_deleting_allowed( ) = abap_true.
+  DELETE FROM zamp_store.
+ELSE.
+  MESSAGE 'no authority' TYPE 'E'.
+ENDIF.
